# model/openapi.yml
openapi: 3.1.0
info:
  title: fleet-server API
  version: "0000-00-00"
  license:
    name: Elastic License 2.0
    url: https://www.elastic.co/licensing/elastic-license
  description: |
    The fleet-server API that is used by agents when enrolled with fleet.

    Note that the current implementations in the fleet-server and elastic-agent may have some difference specifically when it comes to some objects.
    This is most notable when comparing the `Action` implementations. Fleet-server uses a general template for all actions and the elastic-agent will have more specific representations.

    The implementation of fleet-server by default also includes a connection count limiter, as well as limiters for request body sizes.
    If an agent attempts to make request but there are no remaining connections, the attempt will be blocked and the agent will get an error.
    If an agent tries to send a body that is too large the fleet-server will respond with a 400 status code.
components:
  headers:
    apiVersion:
      description: The API version the server is using to respond.
      schema:
        type: string
      examples:
        "2023-06-01":
          value: 2023-06-01
    requestID:
      description: The X-Request-Id header used for tracing requests.
      schema:
        type: string
  securitySchemes:
    apiKey:
      description: API key security will check that the API key exists and is enabled, but will not check additional permissions
      type: apiKey
      in: header
      name: ApiKey
    agentApiKey:
      description: Agent API key security will check that the API key exists, is enabled, and is assigned to the agent
      type: apiKey
      in: header
      name: ApiKey
  schemas:
...
    enrollRequest:
      description: A request to enroll a new agent into fleet.
      type: object
      required:
        - type
        - metadata
      properties:
        id:
          type: string
          description: |
            The ID of the agent.
            This is the ID that will be used to reference this agent, if no ID is passed one will be generated.
            If another agent is enrolled with the same ID the other agent will no longer be able to communicate,
            this new agent is considered a replacement of the other agent. The other agent will be able to continue
            sending data to ES.
        type:
          description: |
            The enrollment type of the agent.
            The agent only supports the PERMANENT value.
            In the future the enrollment type may be used to indicate agents that use fleet for reporting and monitoring, but do not use policies.
          type: string
          enum:
            - PERMANENT
        replace_token:
          type: string
          description: |
            The replacement token of the agent.
            Provided when an agent could replace an existing agent. This token must match the original enrollment of
            that agent otherwise it will not be able to enroll.
        enrollment_id:
          type: string
          description: |
            The enrollment ID of the agent.
            To replace an agent on enroll fail.
            The existing agent with a matching enrollment_id will be deleted if it never checked in. The new agent will be enrolled with the enrollment_id.
        shared_id:
          deprecated: true
          type: string
          description: |
            The shared ID of the agent.
            To support pre-existing installs.

            Never implemented.
        metadata:
          $ref: "#/components/schemas/enrollMetadata"
...
    enrollResponse:
      description: The enrollment action response.
      type: object
      required:
        - action
        - item
      properties:
        action:
          description: The action result. Will have the value "created".
          type: string
        item:
          $ref: "#/components/schemas/enrollResponseItem"
...
  parameters:
    requestId:
      name: X-Request-Id
      description: The request tracking ID for APM.
      in: header
      schema:
        type: string
    apiVersion:
      name: elastic-api-version
      description: The API version to use, format should be "YYYY-MM-DD"
      in: header
      schema:
        type: string
      examples:
        "2023-06-01":
          value: 2023-06-01
    userAgent:
      name: User-Agent
      description: |
        The user-agent header that is sent.
        Must have the format "elastic agent X.Y.Z" where "X.Y.Z" indicates the agent version.
        The agent version must not be greater than the version of the fleet-server.
      in: header
      required: true
      schema:
        type: string
      examples:
        valid:
          description: A valid User-Agent header from the elastic-agent.
          value: elastic agent 8.6.0
        validWithSuffix:
          description: A version number may include an optional suffix
          value: elastic agent 8.6.0-SNAPSHOT
        invalidName:
          description: The elastic-agent name is not formatted correctly.
          value: elastic-agent 8.6.0
        outdatedVersion:
          description: The version string given is too old.
          value: elastic agent 7.0.0
  responses:
...
    keyNotEnabled:
      description: 401 response when the API key is not enabled on any endpoint except /api/fleet/status. Or when there are issues updating an inactive agent on the ack endpoint.
      headers:
        Elastic-Api-Version:
          $ref: "#/components/headers/apiVersion"
        X-Request-Id:
          $ref: "#/components/headers/requestID"
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/error"
          examples:
            unauthorized:
              description: The ApiKey is not enabled
              value:
                statusCode: 401
                error: Unauthorized
                message: ApiKey not enabled
...
paths:
...
  /api/fleet/agents/enroll:
    post:
      operationId: agentEnroll
      parameters:
        - $ref: "#/components/parameters/userAgent"
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      security:
        - apiKey: []
      description: Enroll a new agent to fleet-server. The agent is enrolled in the policy encoded in the apiKey used.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/enrollRequest"
            examples:
              request:
                description: A request to enroll a new agent.
                value:
                  type: PERMANENT
                  metadata:
                    local:
                      elastic:
                        agent:
                          id: ""
                          version: 8.6.0
                          snapshot: false
                      host:
                        hostname: test
                    tags:
                      - us-west
                      - test
      responses:
        "200":
          description: Agent enrolled successfully.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/enrollResponse"
              examples:
                success:
                  description: Agent enrolled successfully.
                  value:
                    action: created
                    item:
                      id: agent-test-id
                      active: true
                      policy_id: agent-policy-id
                      type: PERMANENT
                      enrolled_at: 2022-12-01T01:02:03Z
                      user_provided_metadata: {}
                      local_metadata:
                        elastic:
                          agent:
                            id: agent-test-id
                            version: 8.6.0
                            snapshot: false
                        host:
                          hostname: test
                      access_api_key_id: api-key-id
                      access_api_key: api-key-token
                      status: online
                      tags:
                        - us-west
                        - test
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "408":
          $ref: "#/components/responses/deadline"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/agents/{id}/checkin:
    post:
      operationId: agentCheckin
      description: |
        The agent checkin endpoint.
        Clients will long-poll this endpoint.
        A client may inform fleet-server of it's long-poll timeout in the request body.
        The fleet-server will return a response if there is a new action for the agent, or if the polling timeout is reached.
        The fleet-server may also use some jitter to offset the polling timeout, if specified a random amount of the jitter value may be subtracted from the polling timeout.
        The fleet-sever polling timeout is short-circuited in cases of heavy load where setting up the checkin (ensuring the API key is authed etc) takes longer then the timeout value.
        Fleet-server sets the poll timeout to 5m by default (with a 10m write timeout), for these values we assume that elastic-agent's request timeout is set to 10m and the cloud-proxy's timeout is longer than 10m.
      parameters:
        - name: id
          in: path
          description: The agent ID.
          required: true
          schema:
            type: string
        - name: Accept-Encoding
          in: header
          description: |
            If the agent is able to accept encoded responses.
            Used to indicate if GZIP compression may be used by the server.
            The elastic-agent does not use the accept-encoding header.
          schema:
            type: string
        - $ref: "#/components/parameters/userAgent"
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      security:
        - agentApiKey: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/checkinRequest"
            examples:
              request:
                description: A checkin request from an elastic-agent.
                value:
                  status: online
                  message: ""
                  ack_token: previous-token
              withCheckinDetails:
                description: A checking request from an elastic-agent that is downloading a new version.
                value:
                  status: online
                  message: ""
                  ack_token: previous-token
                  upgrade_details:
                    action_id: upgrade-action-id
                    target_version: X.Y.Z
                    state: UPG_DOWNLOADING
                    metadata:
                      download_percent: 12.3
                      download_rate: 1024
      responses:
        "200":
          description: Agent checkin successful. May include actions.
          headers:
            Content-Encoding:
              description: Responses may be compressed if the accept encoding indicates it. Currently not used by the agent.
              schema:
                type: string
              examples:
                gzip:
                  description: Response is gzip encoded as the request headers allowed it.
                  value: gzip
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/checkinResponse"
              examples:
                response:
                  description: Agent successfully checked in.
                  value:
                    action: checkin
                    ack_token: new-token
                    actions:
                      - agent_id: test-agent
                        created_at: 2022-12-01T01:02:03Z
                        data:
                          log_level: debug
                        id: test-action
                        type: SETTINGS
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/agentNotFound"
        "408":
          $ref: "#/components/responses/deadline"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/agents/{id}/acks:
    post:
      operationId: agentAcks
      description: |
        The endpoint that an agent uses to acknowledge (and inform fleet-server) of events that it has recieved/executed.
        A single action may have multiple different events associated with it.
        Also an action may not have any acks associated with it.

        Note that using this endpoint for an `UPGRADE` action is deprecated behaviour.
        `UPGRADE` status should use the `upgrade_details` attribute of the checkin request body.
      parameters:
        - name: id
          in: path
          description: The agent ID.
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      security:
        - agentApiKey: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ackRequest"
            examples:
              single:
                description: Send an ack for a single event.
                value:
                  events:
                    - type: ACTION_RESULT
                      subtype: ACKNOWLEDGED
                      agent_id: test-agent
                      action_id: test-action
                      timestamp: 2022-12-01T01:02:03.00004-07:00
                      message: Action 'test-action' of type 'TEST' acknowledged.
      responses:
        "200":
          description: Agent ack successfully received.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ackResponse"
              examples:
                single:
                  description: Successfull ack for a single action.
                  value:
                    action: acks
                    errors: false
                    items:
                      - status: 200
                        message: ok
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/agentNotFound"
        "408":
          $ref: "#/components/responses/deadline"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/artifacts/{id}/{sha2}:
    get:
      operationId: artifact
      description: The route to retrieve an artifact from Elasticsearch.
      parameters:
        - name: id
          in: path
          description: The artifact record ID.
          required: true
          schema:
            type: string
        - name: sha2
          in: path
          description: The decoded Sha256 associated with the artifact record.
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      security:
        - agentApiKey: []
      responses:
        "200":
          description: The artifact retrieved from ES.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "404":
          $ref: "#/components/responses/agentNotFound"
        "408":
          $ref: "#/components/responses/deadline"
        "428":
          $ref: "#/components/responses/throttle"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/uploads:
    post:
      operationId: uploadBegin
      summary: Initiate a file upload process
      description: ""
      security:
        - agentApiKey: []
      parameters:
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      requestBody:
        required: true
        description: Information about the file to be uploaded. Minimum required fields are marked as required. Additional fields may be specified and are allowed. For information about the file itself, ECS.file paths are recommended. For archived files, information about the archive should be placed in `file`. Information about the archive members may be placed in a `contents` array.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/uploadBeginRequest"
      responses:
        "200":
          description: Information about the upload procedure
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uploadBeginResponse"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "403":
          $ref: "#/components/responses/forbidden"
        "408":
          $ref: "#/components/responses/deadline"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/uploads/{id}/{chunkNum}:
    put:
      operationId: uploadChunk
      summary: Upload a section of file data
      description: "Upload portions of the intended file in a piecewise fashion. Chunks may be uploaded in any order, and may be uploaded in parallel. The body is the raw contents of the file at the given position matching the chunk number. The body must be the exact chunk size returned from the upload initiation response. All chunks must be this size except for the final one, which is naturally the file remainder."
      security:
        - agentApiKey: []
      parameters:
        - name: id
          in: path
          description: The upload_id as returned in the Upload initiation response
          required: true
          schema:
            type: string
            examples:
              - ecb30383-6dd1-4b1d-bed0-2386b4e5df51
        - name: chunkNum
          in: path
          description: the positional index of the chunk within the file. The first chunk is 0, the next 1, etc.
          required: true
          schema:
            type: integer
            examples:
              - 3
        - name: X-Chunk-SHA2
          in: header
          required: true
          description: the SHA256 hash of the body contents for this request
          schema:
            type: string
            examples:
              - 0c4a81b85a6b7ff00bde6c32e1e8be33b4b793b3b7b5cb03db93f77f7c9374d1
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      requestBody:
        description: The chunk contents as bytes
        required: true
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Successful chunk upload
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "403":
          $ref: "#/components/responses/forbidden"
        "408":
          $ref: "#/components/responses/deadline"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/uploads/{id}:
    post:
      operationId: uploadComplete
      summary: Complete a file upload process
      description: ""
      security:
        - agentApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/uploadCompleteRequest"
      parameters:
        - name: id
          in: path
          description: The upload_id as returned in the Upload initiation response
          required: true
          schema:
            type: string
            examples:
              - ecb30383-6dd1-4b1d-bed0-2386b4e5df51
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: success response
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    examples:
                      - "ok"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "403":
          $ref: "#/components/responses/forbidden"
        "408":
          $ref: "#/components/responses/deadline"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/fleet/file/{id}:
    get:
      operationId: getFile
      summary: retrieve stored file for integration
      description: "Stream out file contents to an agent or integration, provided there is a matching and authorized file stored in elasticsearch."
      security:
        - agentApiKey: []
      parameters:
        - name: id
          in: path
          description: The file_id as provided to the integration
          required: true
          schema:
            type: string
            examples:
              - ecb30383-6dd1-4b1d-bed0-2386b4e5df51
        - $ref: "#/components/parameters/apiVersion"
        - $ref: "#/components/parameters/requestId"
      responses:
        "200":
          description: File Contents.
          headers:
            X-File-Sha2:
              description: SHA256 digest of the file contents. Only sent when the uploaded file contains a file.hash.sha256 value.
              schema:
                type: string
                examples:
                  - 0c4a81b85a6b7ff00bde6c32e1e8be33b4b793b3b7b5cb03db93f77f7c9374d1
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "403":
          description: The requesting entity is not permitted to access the requested file
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                unauthorized:
                  description: The file is not accessible
                  value:
                    statusCode: 403
                    error: Forbidden
                    message: Client is not authorized
        "404":
          description: bad path, file not found
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"
  /api/agents/upgrades/{major}.{minor}.{patch}/pgp-public-key:
    get:
      operationId: getPGPKey
      summary: retrieve a PGP key from the fleet-server's local storage.
      description: "Get a PGP key that can be used to verify agent upgrades. Key is stored on (fleet-server's) disk."
      parameters:
        - name: major
          in: path
          description: The major version number.
          required: true
          schema:
            type: integer
        - name: minor
          in: path
          description: The minor version number.
          required: true
          schema:
            type: integer
        - name: patch
          in: path
          description: The patch version number.
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/apiVersion"
        - $ref: "#/components/parameters/requestId"
      responses:
        "200":
          description: The PGP key bytes.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
          content:
            "application/octet-stream":
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/badRequest"
        "500":
          description: The server has an error retrieving or reading the local key.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
        "501":
          description: The server will not serve the request without a TLS connection.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
  /api/fleet/agents/{id}/audit/unenroll:
    post:
      operationId: auditUnenroll
      summary: Annotate an agent with unenroll attributes.
      description: |
        Annotate an agent with unenroll attributes so the UI will show it as uninstalled or orphaned instead of offline.
        API keys for an agent that is annotated will not be revoked by fleet-server.
        If an agent checks in after the unenroll attributes are set, the attributes will be removed from the agent document.
      security:
        - agentApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auditUnenrollRequest"
            examples:
              request:
                description: An uninstall request.
                value:
                  reason: uninstall
                  timestamp: 2024-01-01T12:00:00.000Z
      parameters:
        - name: id
          in: path
          description: The agent ID.
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/requestId"
        - $ref: "#/components/parameters/apiVersion"
      responses:
        "200":
          description: Audit attributes set in agent document.
          headers:
            Elastic-Api-Version:
              $ref: "#/components/headers/apiVersion"
            X-Request-Id:
              $ref: "#/components/headers/requestID"
        "400":
          $ref: "#/components/responses/badRequest"
        "401":
          $ref: "#/components/responses/keyNotEnabled"
        "409":
          $ref: "#/components/responses/conflict"
        "500":
          $ref: "#/components/responses/internalServerError"
        "503":
          $ref: "#/components/responses/unavailable"